Move test "infrastructure" types
to a Infrastructure folder and namespace
