+++ tests/LibUsbSharp.Tests/Given_a_supported_Huddly_USB_device.cs
@@ -1,7 +1,7 @@
 ﻿using System.Text;
 using FluentAssertions;
 using LibUsbSharp.Descriptor;
-using LibUsbSharp.Tests.TestLogger;
+using LibUsbSharp.Tests.Infrastructure;
 using Microsoft.Extensions.Logging;
 using Xunit;
 using Xunit.Abstractions;
