+++ tests/LibUsbSharp.Tests/Given_any_USB_device.cs
@@ -1,6 +1,6 @@
 using FluentAssertions;
 using LibUsbSharp.Descriptor;
-using LibUsbSharp.Tests.TestLogger;
+using LibUsbSharp.Tests.Infrastructure;
 using Microsoft.Extensions.Logging;
 using Xunit;
 using Xunit.Abstractions;
