+++ tests/LibUsbSharp.Tests/Infrastructure/TestLogger.cs
@@ -1,7 +1,7 @@
 using Microsoft.Extensions.Logging;
 using Xunit.Abstractions;
 
-namespace LibUsbSharp.Tests.TestLogger;
+namespace LibUsbSharp.Tests.Infrastructure;
 
 public class TestLogger(string _categoryName, ITestOutputHelper _output) : ILogger
 {
